requirements.txt
pandas==1.3.5
google-api-python-client
google-cloud-bigquery
google-auth
oauth2client
# plotly
pyarrow
xlsxwriter
numpy==1.22.0